Fix parsing errors and changes
The most relevant change is that for the new challenge a new email is used.
The ACME accounts are bound to the URL (including IP:PORT). So when changing
the endpoint to a separate port number the account is missing. LEGO finds the
account based on the email and then there is also a key mismatch (if transiting
to PQC). Using a new email forces a new account creation in an easier manner.
Other fixes:

* Dates (notBefore and NotAfter) were out of RFC3339 format.
* Change messages in the log
* Parsing port numbers: 10001 to 14000 (there's a URL prefix verification in Pebble,
it throws an error if the accounts prefix URL does not match).

Note: tls.VerifyClientCertIfGiven is being used (for testing http requests/responses).
Should change this later.