Switch between two CA objects
Backup the original CA (e.g., with classical crypto. certs) so if
testing the new challenge (PQC CA) we keep the original CA as a
backup. So, to test the new challenge, the first certificate
issuance is from the classical one, then testing the new challenge
will be using the PQC one; a new test will back up the classical
CA without needing to terminate Pebble.